recreated

+++ SchoolManagementSystem/Controllers/StudentTablesController.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
-using System.Data.Entity.Validation;
 using System.Linq;
 using System.Net;
 using System.Web;
@@ -22,7 +21,7 @@ namespace SchoolManagementSystem.Controllers
             {
                 return RedirectToAction("Login", "Home");
             }
-            var studentTables = db.StudentTables.Include(s => s.ClassTable).Include(s => s.ProgrameTable).Include(s => s.SessionTable).Include(s => s.UserTable).OrderByDescending(s=>s.StudentID);
+            var studentTables = db.StudentTables.Include(s => s.ClassTable).Include(s => s.ProgrameTable).Include(s => s.SessionTable).Include(s => s.UserTable).OrderByDescending(s => s.StudentID); ;
             return View(studentTables.ToList());
         }
 
@@ -52,10 +51,6 @@ namespace SchoolManagementSystem.Controllers
             {
                 return RedirectToAction("Login", "Home");
             }
-            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
-            {
-                return RedirectToAction("Login", "Home");
-            }
             ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name");
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name");
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name");
@@ -76,43 +71,26 @@ namespace SchoolManagementSystem.Controllers
             }
             int userid = Convert.ToInt32(Convert.ToString(Session["UserID"]));
             studentTable.UserID = userid;
-            
-            try
+            if (ModelState.IsValid)
             {
-               if (ModelState.IsValid)
+                db.StudentTables.Add(studentTable);
+                db.SaveChanges();
+                if (studentTable.PhotoFile != null)
                 {
-                    db.StudentTables.Add(studentTable);
-                    db.SaveChanges();
-                    if (studentTable.PhotoFile != null)
+                    var folder = "/Content/StudentPhotos";
+                    var file = string.Format("{0}.png", studentTable.StudentID);
+                    var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
+                    if (response)
                     {
-                        var folder = "/Content/StudentPhotos";
-                        var file = string.Format("{0}.png", studentTable.StudentID);
-                        var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
-                        if (response)
-                        {
-                            var pic = string.Format("{0}/{1}", folder, file);
-                            studentTable.Photo = pic;
-                            db.Entry(studentTable).State = EntityState.Modified;
-                            db.SaveChanges();
-                        }
-                    }
-                    return RedirectToAction("Index");
-                }
-            }
-            catch (DbEntityValidationException e)
-            {
-                foreach (var eve in e.EntityValidationErrors)
-                {
-                    foreach (var ve in eve.ValidationErrors)
-                    {
-                        // Add validation errors to the ModelState
-                        ModelState.AddModelError(ve.PropertyName, ve.ErrorMessage);
+                        var pic = string.Format("{0}/{1}", folder, file);
+                        studentTable.Photo = pic;
+                        db.Entry(studentTable).State = EntityState.Modified;
+                        db.SaveChanges();
                     }
                 }
-                // Handle the error or return to the view with validation errors
-                return View();
+                return RedirectToAction("Index");
             }
-            ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name");
+
             ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name", studentTable.ClassID);
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name", studentTable.ProgrameID);
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name", studentTable.SessionID);
@@ -148,7 +126,7 @@ namespace SchoolManagementSystem.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(StudentTable studentTable)
+        public ActionResult Edit([Bind(Include = "StudentID,SessionID,ProgrameID,ClassID,UserID,Name,FatherName,DateofBirth,Gender,ContactNo,CNIC,FNIC,Photo,AddmissionDate,PreviousSchool,PreviousPercentage,EmailAddress,Address,Nationality,Religion,TribeorCaste,FatherGuardiansOccupation,FatherGuardiansPostalAddress,PhoneOffice,PhoneResident")] StudentTable studentTable)
         {
             if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
             {
@@ -156,23 +134,48 @@ namespace SchoolManagementSystem.Controllers
             }
             int userid = Convert.ToInt32(Convert.ToString(Session["UserID"]));
             studentTable.UserID = userid;
-            if (ModelState.IsValid)
-            {
-                if (studentTable.PhotoFile != null)
+            
+                if (ModelState.IsValid)
                 {
-                    var folder = "/Content/StudentPhotos";
-                    var file = string.Format("{0}.png", studentTable.StudentID);
-                    var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
-                    if (response)
+                    try
                     {
-                        var pic = string.Format("{0}/{1}", folder, file);
-                        studentTable.Photo = pic;
+                        db.Entry(studentTable).State = EntityState.Modified;
+                        db.SaveChanges();
+                        if (studentTable.PhotoFile != null)
+                        {
+                            var folder = "/Content/StudentPhotos";
+                            var file = string.Format("{0}.png", studentTable.StudentID);
+                            var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
+                            if (response)
+                            {
+                                var pic = string.Format("{0}/{1}", folder, file);
+                                studentTable.Photo = pic;
+                                db.Entry(studentTable).State = EntityState.Modified;
+                                db.SaveChanges();
+                            }
+                        }
+                        return RedirectToAction("Index");
                     }
+                    catch (System.Data.Entity.Validation.DbEntityValidationException dbEx)
+                                {
+                                    Exception raise = dbEx;
+                                    foreach (var validationErrors in dbEx.EntityValidationErrors)
+                                    {
+                                        foreach (var validationError in validationErrors.ValidationErrors)
+                                        {
+                                            string message = string.Format("{0}:{1}",
+                                                validationErrors.Entry.Entity.ToString(),
+                                                validationError.ErrorMessage);
+                                            // raise a new exception nesting  
+                                            // the current instance as InnerException  
+                                            raise = new InvalidOperationException(message, raise);
+                                        }
+                                    }
+                                    throw raise;
+                                }
                 }
-                db.Entry(studentTable).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+            
+            
             ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name", studentTable.ClassID);
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name", studentTable.ProgrameID);
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name", studentTable.SessionID);
@@ -183,6 +186,10 @@ namespace SchoolManagementSystem.Controllers
         // GET: StudentTables/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -200,6 +207,10 @@ namespace SchoolManagementSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             StudentTable studentTable = db.StudentTables.Find(id);
             db.StudentTables.Remove(studentTable);
             db.SaveChanges();
